Refactor CLI auth into a dedicated lib/auth layer
Extract the OAuth device-login logic that landed in commands/generic.ts
into a cohesive lib/auth/ layer, preserving all behavior:

- lib/auth/oauth.ts: OAuth2 client factory, device-token polling, token
  refresh, revoke, and id_token decoding (getValidAccessToken moved here
  from sdks.ts; the 3x-duplicated Oauth2 client construction is unified
  behind createOauth2).
- lib/auth/session.ts: typed session accessor, classification + current-
  session helpers, deleteServerSession, and a single logoutSessions() that
  replaces the three near-identical cleanup loops.
- lib/auth/login.ts: login orchestration and flows (loginCommand,
  loginWithEmailPassword, loginWithOAuthDevice, switchToAccount,
  completeMfaLogin, getCurrentAccount).
- utils.ts: endpoint classifiers placed next to isCloudHostname.
- generic.ts slimmed from ~1060 to ~330 lines (command definitions only).

Repointed sdks.ts/deployment.ts imports and registered the new auth
files in CLI.php getFiles(). Device-login output uses process.stdout.write
(matching the existing spinner idiom); dropped thin wrapper helpers.
