+++ templates/cli/lib/commands/generic.ts
@@ -1,12 +1,7 @@
 import inquirer from "inquirer";
 import { Command } from "commander";
-import { AppwriteException, Client } from "@appwrite.io/console";
-import { OAUTH2_CLIENT_ID, OAUTH2_SCOPES, sdkForConsole } from "../sdks.js";
-import {
-  globalConfig,
-  localConfig,
-  normalizeCloudConsoleEndpoint,
-} from "../config.js";
+import { Client } from "@appwrite.io/console";
+import { globalConfig, localConfig } from "../config.js";
 import { EXECUTABLE_NAME } from "../constants.js";
 import {
   actionRunner,
@@ -16,711 +11,27 @@ import {
   error,
   parse,
   hint,
-  warn,
   log,
   drawTable,
   cliConfig,
 } from "../parser.js";
-import { isCloudHostname } from "../utils.js";
 import ID from "../id.js";
+import { questionsLogout } from "../questions.js";
+import { getCurrentAccount, loginCommand } from "../auth/login.js";
 import {
-  questionsListFactors,
-  questionsLogin,
-  questionsLogout,
-  questionsMFAChallenge,
-  questionsSwitchAccount,
-} from "../questions.js";
-import { Account, Oauth2, type Models } from "@appwrite.io/console";
-import ClientLegacy from "../client.js";
-
-const DEFAULT_ENDPOINT = "https://cloud.appwrite.io/v1";
-
-interface AppwriteError {
-  type?: string;
-  response?: string;
-}
-
-const isGuestUnauthorizedError = (err: unknown): err is AppwriteError =>
-  (err as AppwriteError)?.type === "general_unauthorized_scope" ||
-  (err as AppwriteError)?.response === "general_unauthorized_scope";
-
-const isMfaRequiredError = (err: unknown): err is AppwriteError =>
-  (err as AppwriteError)?.type === "user_more_factors_required" ||
-  (err as AppwriteError)?.response === "user_more_factors_required";
-
-const isAuthorizationPendingError = (err: unknown): boolean => {
-  if (!(err instanceof AppwriteException)) {
-    return false;
-  }
-
-  const response = typeof err.response === "string" ? err.response : "";
-
-  return (
-    err.type === "authorization_pending" ||
-    err.type === "slow_down" ||
-    err.message === "authorization_pending" ||
-    err.message === "slow_down" ||
-    response.includes("authorization_pending") ||
-    response.includes("slow_down")
-  );
-};
-
-const decodeIdToken = (idToken: string): { email?: string; name?: string; sub?: string } => {
-  try {
-    const payload = idToken.split(".")[1];
-    if (!payload) return {};
-    const decoded = JSON.parse(Buffer.from(payload, "base64url").toString("utf-8"));
-    return {
-      email: decoded.email,
-      name: decoded.name,
-      sub: decoded.sub,
-    };
-  } catch (_error) {
-    return {};
-  }
-};
-
-const isRegionalCloudEndpoint = (endpoint: string): boolean => {
-  try {
-    const hostname = new URL(endpoint).hostname;
-    return isCloudHostname(hostname) && hostname !== "cloud.appwrite.io";
-  } catch (_error) {
-    return false;
-  }
-};
-
-const isLocalhostHostname = (hostname: string): boolean =>
-  hostname === "localhost" || hostname === "127.0.0.1" || hostname === "[::1]";
-
-const isDevCloudLoginOverrideEnabled = (): boolean =>
-  ["1", "true", "yes"].includes(
-    (process.env.APPWRITE_CLI_DEV_CLOUD_LOGIN ?? "").toLowerCase(),
-  );
-
-const isCloudLoginEndpoint = (endpoint: string): boolean => {
-  try {
-    const hostname = new URL(endpoint).hostname;
-    return (
-      isCloudHostname(hostname) ||
-      (isDevCloudLoginOverrideEnabled() && isLocalhostHostname(hostname))
-    );
-  } catch (_error) {
-    return false;
-  }
-};
-
-const restoreCurrentSession = (sessionId: string): void => {
-  globalConfig.setCurrentSession(
-    globalConfig.getSessionIds().includes(sessionId) ? sessionId : "",
-  );
-};
-
-const restoreCurrentSessionFallback = (
-  preferredSessionId: string,
-  fallbackSessionIds: string[],
-): void => {
-  const sessionIds = globalConfig.getSessionIds();
-  globalConfig.setCurrentSession(
-    [preferredSessionId, ...fallbackSessionIds].find((sessionId) =>
-      sessionIds.includes(sessionId),
-    ) ?? "",
-  );
-};
-
-const removeCurrentSession = (): void => {
-  const current = globalConfig.getCurrentSession();
-  globalConfig.setCurrentSession("");
-  globalConfig.removeSession(current);
-};
-
-const hasAuthSession = (): boolean => {
-  return globalConfig.getAccessToken() !== "" || globalConfig.getCookie() !== "";
-};
-
-const getCurrentAccount = async (): Promise<Models.User | null> => {
-  if (globalConfig.getEndpoint() === "" || !hasAuthSession()) {
-    return null;
-  }
-
-  const endpoint = normalizeCloudConsoleEndpoint(globalConfig.getEndpoint());
-  if (endpoint !== globalConfig.getEndpoint()) {
-    globalConfig.setEndpoint(endpoint);
-  }
-
-  const client = await sdkForConsole();
-  const accountClient = new Account(client);
-
-  try {
-    const account = await accountClient.get();
-    globalConfig.setEmail(account.email);
-    return account;
-  } catch (err) {
-    if (isGuestUnauthorizedError(err)) {
-      removeCurrentSession();
-    }
-    return null;
-  }
-};
-
-const createLegacyConsoleClient = (endpoint: string): ClientLegacy => {
-  const legacyClient = new ClientLegacy();
-  legacyClient.setEndpoint(endpoint);
-  legacyClient.setProject("console");
-  if (globalConfig.getSelfSigned()) {
-    legacyClient.setSelfSigned(true);
-  }
-  return legacyClient;
-};
-
-const completeMfaLogin = async ({
-  client,
-  legacyClient,
-  mfa,
-  code,
-}: {
-  client: Client;
-  legacyClient: ClientLegacy;
-  mfa?: string;
-  code?: string;
-}): Promise<Models.User> => {
-  let accountClient = new Account(client);
-
-  const savedCookie = globalConfig.getCookie();
-  if (savedCookie) {
-    legacyClient.setCookie(savedCookie);
-    client.setCookie(savedCookie);
-  }
-
-  const { factor } = mfa
-    ? { factor: mfa }
-    : await inquirer.prompt(questionsListFactors);
-  const challenge = await accountClient.createMfaChallenge(factor);
-
-  const { otp } = code
-    ? { otp: code }
-    : await inquirer.prompt(questionsMFAChallenge);
-  await legacyClient.call(
-    "PUT",
-    "/account/mfa/challenges",
-    {
-      "content-type": "application/json",
-    },
-    {
-      challengeId: challenge.$id,
-      otp,
-    },
-  );
-
-  const updatedCookie = globalConfig.getCookie();
-  if (updatedCookie) {
-    client.setCookie(updatedCookie);
-  }
-
-  accountClient = new Account(client);
-  return accountClient.get();
-};
-
-const loginWithEmailPassword = async ({
-  id,
-  oldCurrent,
-  configEndpoint,
-  email,
-  password,
-  endpoint,
-  mfa,
-  code,
-}: {
-  id: string;
-  oldCurrent: string;
-  configEndpoint: string;
-  email: string;
-  password: string;
-  endpoint?: string;
-  mfa?: string;
-  code?: string;
-}): Promise<void> => {
-  globalConfig.addSession(id, { endpoint: configEndpoint });
-  globalConfig.setCurrentSession(id);
-  globalConfig.setEndpoint(configEndpoint);
-  globalConfig.setEmail(email);
-
-  const legacyClient = createLegacyConsoleClient(configEndpoint);
-  const client = await sdkForConsole({ requiresAuth: false });
-  let accountClient = new Account(client);
-  let account: Models.User;
-
-  try {
-    await legacyClient.call(
-      "POST",
-      "/account/sessions/email",
-      {
-        "content-type": "application/json",
-      },
-      { email, password },
-    );
-
-    const savedCookie = globalConfig.getCookie();
-    if (savedCookie) {
-      legacyClient.setCookie(savedCookie);
-      client.setCookie(savedCookie);
-    }
-
-    accountClient = new Account(client);
-    account = await accountClient.get();
-  } catch (err) {
-    if (isMfaRequiredError(err)) {
-      try {
-        account = await completeMfaLogin({
-          client,
-          legacyClient,
-          mfa,
-          code,
-        });
-      } catch (mfaErr) {
-        globalConfig.removeSession(id);
-        restoreCurrentSession(oldCurrent);
-        throw mfaErr;
-      }
-    } else {
-      globalConfig.removeSession(id);
-      globalConfig.setCurrentSession(oldCurrent);
-
-      if (
-        endpoint !== DEFAULT_ENDPOINT &&
-        ((err as AppwriteError)?.type === "user_invalid_credentials" ||
-          (err as AppwriteError)?.response === "user_invalid_credentials")
-      ) {
-        log("Use the --endpoint option for self-hosted instances");
-      }
-
-      throw err;
-    }
-  }
-
-  globalConfig.setEmail(account.email);
-  success("Successfully signed in as " + account.email);
-  hint(
-    "Next you can create or link to your project using 'appwrite init project'",
-  );
-};
-
-const deleteServerSession = async (sessionId: string): Promise<boolean> => {
-  const session = globalConfig.get(sessionId) as
-    | {
-        refreshToken?: string;
-        endpoint?: string;
-        clientId?: string;
-        cookie?: string;
-      }
-    | undefined;
-  if (!session?.endpoint) {
-    return false;
-  }
-
-  try {
-    if (session.refreshToken) {
-      const client = new Client()
-        .setEndpoint(session.endpoint)
-        .setProject("console")
-        .setSelfSigned(globalConfig.getSelfSigned());
-      const oauth2 = new Oauth2(client);
-      await oauth2.revoke({
-        projectId: "console",
-        token: session.refreshToken,
-        tokenTypeHint: "refresh_token",
-        clientId: session.clientId || OAUTH2_CLIENT_ID,
-      });
-      return true;
-    }
-
-    if (session.cookie) {
-      const legacyClient = createLegacyConsoleClient(session.endpoint);
-      legacyClient.setCookie(session.cookie);
-      await legacyClient.call(
-        "DELETE",
-        "/account/sessions/current",
-        {
-          "content-type": "application/json",
-        },
-      );
-      return true;
-    }
-
-    return false;
-  } catch (_e) {
-    return false;
-  }
-};
-
-const deleteLocalSession = (accountId: string): void => {
-  globalConfig.removeSession(accountId);
-};
-
-const isLocalOnlySession = (sessionId: string): boolean => {
-  const session = globalConfig.get(sessionId) as
-    | { refreshToken?: string; cookie?: string }
-    | undefined;
-
-  return Boolean(session && !session.refreshToken && !session.cookie);
-};
-
-const getSessionAccountKey = (sessionId: string): string | undefined => {
-  const session = globalConfig.get(sessionId) as
-    | { email?: string; endpoint?: string }
-    | undefined;
-  if (!session) return undefined;
-  return `${session.email ?? ""}|${normalizeCloudConsoleEndpoint(
-    session.endpoint ?? "",
-  )}`;
-};
-
-const sleep = (ms: number): Promise<void> => {
-  return new Promise((resolve) => setTimeout(resolve, ms));
-};
-
-const startWaitingForApprovalSpinner = (): (() => void) => {
-  const message = "Waiting for approval...";
-  const frames = ["⠋", "⠙", "⠹", "⠸", "⠼", "⠴", "⠦", "⠧", "⠇", "⠏"];
-
-  if (!process.stdout.isTTY) {
-    console.log(message);
-    return () => {};
-  }
-
-  let frame = 0;
-  process.stdout.write(`${frames[frame]} ${message}`);
-
-  const interval = setInterval(() => {
-    frame = (frame + 1) % frames.length;
-    process.stdout.write(`\r${frames[frame]} ${message}`);
-  }, 80);
-
-  return () => {
-    clearInterval(interval);
-    process.stdout.write("\r\x1b[K");
-  };
-};
-
-const isLegacySession = (sessionId: string): boolean => {
-  const session = globalConfig.get(sessionId) as
-    | { accessToken?: string; cookie?: string }
-    | undefined;
-
-  return Boolean(session?.cookie && !session?.accessToken);
-};
-
-const removeLegacySessionsExcept = async (
-  sessionIdToKeep: string,
-): Promise<{ removed: number; failed: number }> => {
-  let removed = 0;
-  let failed = 0;
-
-  for (const sessionId of globalConfig.getSessionIds()) {
-    if (sessionId === sessionIdToKeep || !isLegacySession(sessionId)) {
-      continue;
-    }
-
-    const serverDeleted = await deleteServerSession(sessionId);
-    if (serverDeleted) {
-      globalConfig.removeSession(sessionId);
-      removed++;
-    } else {
-      failed++;
-    }
-  }
-
-  return { removed, failed };
-};
-
-/**
- * Given selected session IDs, determine which local sessions belong to the
- * selected accounts and should be individually logged out from the server.
- *
- * @param selectedSessionIds Array of session IDs to process for logout.
- * @returns Session IDs that belong to selected account groups.
- */
-const planSessionLogout = (
-  selectedSessionIds: string[],
-): string[] => {
-  // Map to group all session IDs by their unique account key (email+endpoint)
-  const sessionIdsByAccount = new Map<string, string[]>();
-  for (const sessionId of globalConfig.getSessionIds()) {
-    const key = getSessionAccountKey(sessionId);
-    if (!key) continue; // Skip sessions without proper account key
-
-    // For each account key, gather all associated session IDs
-    const ids = sessionIdsByAccount.get(key) ?? [];
-    ids.push(sessionId);
-    sessionIdsByAccount.set(key, ids);
-  }
-
-  // Map to store one selected session ID per unique account for server logout
-  const selectedByAccount = new Map<string, string>();
-  for (const selectedSessionId of selectedSessionIds) {
-    const key = getSessionAccountKey(selectedSessionId);
-    if (!key || selectedByAccount.has(key)) continue; // Skip if key missing or already considered for this account
-    selectedByAccount.set(key, selectedSessionId);
-  }
-
-  return Array.from(selectedByAccount.keys()).flatMap(
-    (accountKey) => sessionIdsByAccount.get(accountKey) ?? [],
-  );
-};
-
-export const loginCommand = async ({
-  email,
-  password,
-  endpoint,
-  mfa,
-  code,
-  switch: switchAccount,
-  new: newAccount,
-}: {
-  email?: string;
-  password?: string;
-  endpoint?: string;
-  mfa?: string;
-  code?: string;
-  switch?: boolean;
-  new?: boolean;
-}): Promise<void> => {
-  let oldCurrent = globalConfig.getCurrentSession();
-
-  if (switchAccount && newAccount) {
-    throw new Error("Use either --switch or --new, not both.");
-  }
-
-  if (endpoint && isRegionalCloudEndpoint(endpoint)) {
-    throw new Error(
-      `Cloud login uses ${DEFAULT_ENDPOINT}. Regional Cloud endpoints are for project API calls, not account login.`,
-    );
-  }
-
-  const configEndpoint = normalizeCloudConsoleEndpoint(
-    (endpoint ?? globalConfig.getEndpoint()) || DEFAULT_ENDPOINT,
-  );
-  const shouldUseCloudLogin = isCloudLoginEndpoint(configEndpoint);
-
-  oldCurrent = globalConfig.getCurrentSession();
-
-  if (oldCurrent !== "" && !newAccount) {
-    let account: Models.User | null = null;
-    try {
-      account = await getCurrentAccount();
-    } catch (_err) {
-      account = null;
-    }
-    oldCurrent = globalConfig.getCurrentSession();
-
-    if (account) {
-      if (!globalConfig.getAccessToken() && globalConfig.getCookie()) {
-        warn(
-          `You are using a legacy cookie session. Run '${EXECUTABLE_NAME} login --new' to switch to the new browser-based login flow.`,
-        );
-      }
-
-      if (!email && !password && !endpoint && !switchAccount && !newAccount) {
-        success("Already logged in as " + account.email);
-        hint(`Use '${EXECUTABLE_NAME} login --new' to add another account`);
-        return;
-      }
-    }
-  }
-
-  let answers;
-  if (switchAccount) {
-    if (!globalConfig.getSessions().some((session) => session.email)) {
-      throw new Error(
-        `No signed-in accounts found. Run '${EXECUTABLE_NAME} login' to sign in.`,
-      );
-    }
-    answers = await inquirer.prompt(questionsSwitchAccount);
-  } else if (!shouldUseCloudLogin) {
-    answers =
-      email && password
-        ? { email, password }
-        : await inquirer.prompt(questionsLogin);
-  }
-
-  if (switchAccount && answers?.accountId) {
-    const accountId = answers.accountId;
-
-    if (!globalConfig.getSessionIds().includes(accountId)) {
-      throw Error("Session ID not found");
-    }
-
-    if (accountId === oldCurrent) {
-      const account = await getCurrentAccount();
-      if (account) {
-        success(`Already using ${account.email}`);
-        return;
-      }
-      throw new Error(
-        `Selected account session is no longer valid. Run '${EXECUTABLE_NAME} login --switch' again.`,
-      );
-    }
-
-    globalConfig.setCurrentSession(accountId);
-    globalConfig.setEndpoint(
-      normalizeCloudConsoleEndpoint(globalConfig.getEndpoint()),
-    );
-
-    let account: Models.User | null = null;
-    try {
-      account = await getCurrentAccount();
-    } catch (err) {
-      restoreCurrentSession(oldCurrent);
-      throw err;
-    }
-
-    if (!account) {
-      restoreCurrentSession(oldCurrent);
-      throw new Error(
-        `Selected account session is no longer valid. Run '${EXECUTABLE_NAME} login --switch' again.`,
-      );
-    }
-
-    success(`Switched to ${account.email}`);
-    return;
-  }
-
-  const id = ID.unique();
-
-  if (!shouldUseCloudLogin) {
-    await loginWithEmailPassword({
-      id,
-      oldCurrent,
-      configEndpoint,
-      email: answers.email,
-      password: answers.password,
-      endpoint,
-      mfa,
-      code,
-    });
-    return;
-  }
-
-  const clientId = OAUTH2_CLIENT_ID;
-  const oauth2Client = new Client()
-    .setEndpoint(configEndpoint)
-    .setProject("console")
-    .setSelfSigned(globalConfig.getSelfSigned());
-  const oauth2 = new Oauth2(oauth2Client);
-
-  globalConfig.addSession(id, { endpoint: configEndpoint, clientId });
-  globalConfig.setCurrentSession(id);
-  globalConfig.setEndpoint(configEndpoint);
-
-  let deviceAuth;
-  try {
-    deviceAuth = await oauth2.createDeviceAuthorization({
-      clientId,
-      scope: OAUTH2_SCOPES,
-    });
-  } catch (err) {
-    globalConfig.removeSession(id);
-    globalConfig.setCurrentSession(oldCurrent);
-    throw err;
-  }
-
-  console.log("");
-  console.log("To sign in, confirm the code below in your browser:");
-  console.log("");
-  console.log(`  Code: ${deviceAuth.user_code}`);
-  console.log(`  URL:  ${deviceAuth.verification_uri_complete || deviceAuth.verification_uri}`);
-  console.log("");
-
-  const expiresAt = Date.now() + deviceAuth.expires_in * 1000;
-  let token: Models.Oauth2Token | null = null;
-  const stopWaitingForApprovalSpinner = startWaitingForApprovalSpinner();
-
-  try {
-    while (Date.now() < expiresAt) {
-      await sleep(deviceAuth.interval * 1000);
-
-      try {
-        token = await oauth2.createToken({
-          grantType: "urn:ietf:params:oauth:grant-type:device_code",
-          deviceCode: deviceAuth.device_code,
-          clientId,
-        });
-      } catch (err) {
-        if (isAuthorizationPendingError(err)) {
-          continue;
-        }
-
-        globalConfig.removeSession(id);
-        globalConfig.setCurrentSession(oldCurrent);
-        throw err;
-      }
-
-      if (token) {
-        break;
-      }
-    }
-  } finally {
-    stopWaitingForApprovalSpinner();
-  }
-
-  if (!token || !token.access_token) {
-    globalConfig.removeSession(id);
-    globalConfig.setCurrentSession(oldCurrent);
-    throw new Error("Device authorization timed out or was denied.");
-  }
-
-  const tokenExpiry = Date.now() + token.expires_in * 1000;
-  globalConfig.setAccessToken(token.access_token);
-  globalConfig.setRefreshToken(token.refresh_token || "");
-  globalConfig.setTokenExpiry(tokenExpiry);
-
-  let tokenEmail = "";
-  if (token.id_token) {
-    const idTokenClaims = decodeIdToken(token.id_token);
-    tokenEmail = idTokenClaims.email || "";
-  }
-
-  if (tokenEmail) {
-    globalConfig.setEmail(tokenEmail);
-  }
-
-  let account: Models.User | null = null;
-  try {
-    account = await getCurrentAccount();
-    if (!account?.email) {
-      throw new Error("Unable to verify the new session.");
-    }
-  } catch (err) {
-    await deleteServerSession(id);
-    globalConfig.removeSession(id);
-    restoreCurrentSession(oldCurrent);
-    throw err;
-  }
-
-  globalConfig.setEmail(account.email);
-
-  const { removed: removedLegacySessions, failed: failedLegacySessions } =
-    await removeLegacySessionsExcept(id);
-
-  success("Successfully signed in as " + globalConfig.getEmail());
-  if (removedLegacySessions > 0) {
-    hint("Removed legacy cookie session data from this CLI configuration.");
-  }
-  if (failedLegacySessions > 0) {
-    warn("Could not revoke all legacy sessions; kept them in local config.");
-  }
-  hint(
-    "Next you can create or link to your project using 'appwrite init project'",
-  );
-};
+  hasAuthSession,
+  logoutSessions,
+  planSessionLogout,
+  restoreCurrentSessionFallback,
+} from "../auth/session.js";
+
+export { loginCommand };
 
 export const whoami = new Command("whoami")
   .description(commandDescriptions["whoami"])
   .action(
     actionRunner(async () => {
-      if (
-        globalConfig.getEndpoint() === "" ||
-        !hasAuthSession()
-      ) {
+      if (globalConfig.getEndpoint() === "" || !hasAuthSession()) {
         error("No user is signed in. To sign in, run 'appwrite login'");
         return;
       }
@@ -794,29 +105,15 @@ export const logout = new Command("logout")
         return;
       }
       if (sessions.length === 1) {
-        const sessionTargets = planSessionLogout([current]);
-        let failed = 0;
-        const failedSessionIds: string[] = [];
-
-        for (const sessionId of sessionTargets) {
-          if (isLocalOnlySession(sessionId)) {
-            deleteLocalSession(sessionId);
-            continue;
-          }
-
-          globalConfig.setCurrentSession(sessionId);
-          const serverDeleted = await deleteServerSession(sessionId);
-          if (serverDeleted) {
-            deleteLocalSession(sessionId);
-          } else {
-            failed++;
-            failedSessionIds.push(sessionId);
-          }
-        }
+        const { failed, failedIds } = await logoutSessions(
+          planSessionLogout([current]),
+        );
 
         if (failed > 0) {
-          restoreCurrentSessionFallback(originalCurrent, failedSessionIds);
-          hint("Could not reach server for all sessions; kept local session data");
+          restoreCurrentSessionFallback(originalCurrent, failedIds);
+          hint(
+            "Could not reach server for all sessions; kept local session data",
+          );
         } else {
           globalConfig.setCurrentSession("");
         }
@@ -828,28 +125,14 @@ export const logout = new Command("logout")
       const answers = await inquirer.prompt(questionsLogout);
 
       if (answers.accounts?.length) {
-        const sessionTargets = planSessionLogout(
-          answers.accounts as string[],
+        const { failed } = await logoutSessions(
+          planSessionLogout(answers.accounts as string[]),
         );
-        let failed = 0;
-
-        for (const sessionId of sessionTargets) {
-          if (isLocalOnlySession(sessionId)) {
-            deleteLocalSession(sessionId);
-            continue;
-          }
-
-          globalConfig.setCurrentSession(sessionId);
-          const serverDeleted = await deleteServerSession(sessionId);
-          if (serverDeleted) {
-            deleteLocalSession(sessionId);
-          } else {
-            failed++;
-          }
-        }
 
         if (failed > 0) {
-          hint("Could not reach server for all sessions; kept local session data");
+          hint(
+            "Could not reach server for all sessions; kept local session data",
+          );
         }
       }
 
@@ -1005,27 +288,15 @@ export const client = new Command("client")
 
         if (reset !== undefined) {
           const originalCurrent = globalConfig.getCurrentSession();
-          let failed = 0;
-          const failedSessionIds: string[] = [];
-          for (const sessionId of globalConfig.getSessionIds()) {
-            if (isLocalOnlySession(sessionId)) {
-              deleteLocalSession(sessionId);
-              continue;
-            }
-
-            globalConfig.setCurrentSession(sessionId);
-            const serverDeleted = await deleteServerSession(sessionId);
-            if (serverDeleted) {
-              deleteLocalSession(sessionId);
-            } else {
-              failed++;
-              failedSessionIds.push(sessionId);
-            }
-          }
+          const { failed, failedIds } = await logoutSessions(
+            globalConfig.getSessionIds(),
+          );
 
           if (failed > 0) {
-            restoreCurrentSessionFallback(originalCurrent, failedSessionIds);
-            hint("Could not reach server for all sessions; kept local session data");
+            restoreCurrentSessionFallback(originalCurrent, failedIds);
+            hint(
+              "Could not reach server for all sessions; kept local session data",
+            );
           } else {
             globalConfig.setCurrentSession("");
           }
